+++ biscuits/src/main/java/com/agathevaisse/biscuits/infrastructure/kid/KidMapper.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 public class KidMapper implements RowMapper<Kid> {
     public Kid mapRow(ResultSet rs, int rowNumber) throws SQLException {
         Kid kid = new Kid();
-        kid.setId(rs.getInt("kid_id"));
+        kid.setId(rs.getLong("kid_id"));
         kid.setNickname(rs.getString("kid_nickname"));
         kid.setImageURL(rs.getString("kid_image"));
         kid.setBiscuitsEarned(rs.getInt("kid_biscuits"));
